python_dramatiq_example/.dagger/pyproject.toml
[project]
name = "python-dramatiq-example"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = ["dagger-io"]

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[tool.uv.sources]
dagger-io = { path = "sdk", editable = true }
